requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
asgiref==3.8.1
asyncio==3.4.3
attrs==24.2.0
disutils==1.4.32.post2
Django==5.1
django-jalali-date==1.1.3
django-js-asset==2.2.0
frozenlist==1.4.1
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
multidict==6.0.5
setuptools==72.1.0
sqlparse==0.5.1
tzdata==2024.1
yarl==1.9.4
